add: Implementarion os the v1.1.0

+++ mendes-soft-api/database/migrations/1606418351910_address_people.ts
@@ -4,26 +4,27 @@ export default class AddressPeople extends BaseSchema {
   protected tableName = "address_people";
 
   public async up() {
-    this.schema.createTable(this.tableName, (table) => {
-      table.bigIncrements("id");
+    if (!(await this.schema.hasTable(this.tableName)))
+      this.schema.createTable(this.tableName, (table) => {
+        table.bigIncrements("id");
 
-      table
-        .bigInteger("person_id")
-        .references("people.id")
-        .notNullable()
-        .onDelete("cascade");
-      table
-        .bigInteger("address_id")
-        .references("addresses.id")
-        .notNullable()
-        .onDelete("cascade");
-      table.boolean("active").defaultTo(true);
+        table
+          .bigInteger("person_id")
+          .references("people.id")
+          .notNullable()
+          .onDelete("cascade");
+        table
+          .bigInteger("address_id")
+          .references("addresses.id")
+          .notNullable()
+          .onDelete("cascade");
+        table.boolean("active").defaultTo(true);
 
-      table.timestamps(true);
-    });
+        table.timestamps(true);
+      });
   }
 
   public async down() {
-    this.schema.dropTable(this.tableName);
+    // this.schema.dropTable(this.tableName);
   }
 }
